+++ src/main/java/janggi/domain/piece/Implementation/Jang.java
@@ -1,5 +1,7 @@
-package janggi.domain.piece;
+package janggi.domain.piece.Implementation;
 
+import janggi.domain.piece.PieceType;
+import janggi.domain.piece.template.AbstractNormalPiece;
 import janggi.domain.status.Team;
 
 public class Jang extends AbstractNormalPiece {
